README.md
This code can tell you whether the ISS was in sunlight or darkness, based on the latitude, longitude, date and time of the measurement. It takes the required data from Excel and then gives you back a csv file.

You need to download libraries: openpyxl, datetime, suntime, pytz.
It assumes that your Excel file is named 'iss_data.xlsx', the original time is GMT + 1 and the date is set to 20th May 2023. You can change this in the code.

Check that your Excel has the same data format as the code expects.